+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - built-in REPL setting `check_badness` to render a badness panel after each parse, available regardless of the active parser plugin.
 - more REPL commands: /load, /unload, /save, /save-parse, /search, /count, /count-csv, /types, /transform, /classify, /unload-parser, /parse-results, /tag and /tags.
 - parser plugin settings have an optional flag to indicate that no parser reload is nodes, REPL adjusted to accommodate this mechanism.
+- functional pattern `deep`.
 
 ### Changed